adding deleting products and cart working fine

+++ src/main/java/com/bilskik/onlineshop/services/ProductService.java
@@ -4,14 +4,24 @@ import com.bilskik.onlineshop.entities.Product;
 import com.bilskik.onlineshop.entities.ProductCategory;
 import com.bilskik.onlineshop.repositories.ProductCategoryRepository;
 import com.bilskik.onlineshop.repositories.ProductRepository;
+import jakarta.transaction.Transactional;
+import lombok.extern.java.Log;
+import org.apache.coyote.Response;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.sql.SQLException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
 public class ProductService {
+
+    private final static Logger logger = LoggerFactory.getLogger(ProductService.class);
     @Autowired
     private ProductRepository productRepository;
     @Autowired
@@ -20,7 +30,7 @@ public class ProductService {
     public List<Product> getAllProducts() {
         List<Product> list = productRepository.findAll();
         if(list.isEmpty()) {
-            throw new IllegalStateException("There is no product available!");
+            throw new NoSuchElementException("There is no product available!");
         }
         return list;
     }
@@ -37,8 +47,31 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public String deleteProduct(int productId) {
+    public void deleteProduct(int productId) {
+        Optional<Product> product = productRepository.findById(productId);
+        if(product.isEmpty()) {
+            throw new NoSuchElementException("There is no product available in DB!");
+        }
+        product.get().setCart(null);
+        productRepository.save(product.get());
         productRepository.deleteById(productId);
-        return "delete Gituwa!";
+    }
+
+    @Transactional
+    public Product updateProduct(Product product) {
+        int result = productRepository.updateProduct(product.getProductId(), product.getProductName()
+                ,product.getAmount(),product.getPrice(),product.getProductDetails(), product.getProductCategory());
+        if(result == 1) {
+            Optional<Product> productOptional = productRepository.findById(product.getProductId());
+            if(productOptional.isPresent()) {
+                return productOptional.get();
+            }
+            else {
+                throw new NoSuchElementException("Product was not found!");
+            }
+        }
+        else {
+            return product;
+        }
     }
 }
